More Testing done and fixed bug
Fixed a bug (line 55). Did more testing on other images. As expected, it doesn't do well with two sets of staves. Only designed to find one staff. These tests also showed that the program is sensitive to the thickness of the staff lines: a single thick staff line can count for multiple lines. Need to run "vertical_runs.py" to find the thickness of staff lines.