Bugs
Resolved a bug where the thickness of the lines was causing issues finding the start/stop of the stafflines. Going through example music and trying to create solutions to problems. Pictures are very sensitive to thresholding, so will have to find good thresholding. Currently working on a way to handle music with black borders (like example_music_4)