Found undetected pairs of eighth notes
Finally happened. Had to check if notes were close to the middle line, where the author has great flexibility to locate the beam. I get 100% of the eighth notes (and looks like 100% accuracy overall). I think this is as much progress that can be made in the time we have.